Added menu button customization
- The notes tab button is hidden
- The activity tab button is now displayed
- Notebook toggle is now disabled (no longer needed)

I've also bundled the latest version of NickelMenu as per this commit:
https://github.com/pgaskin/NickelMenu/commit/2f7b78f26a162e1f598429e7ac18808db79af111

You may want to get a more up-to-date version for your device if
possible, but I'm including a working copy for the current version
of the firmware that supports navigation customization.

Note that I've also added the documentation to the repo, since the
latest documentation includes instructions on how to tweak the tab
navigation at the bottom of the device.